Switch from hashing whole slices to just hashing their lengths
Since the caches are cleared between each row, and the subslices used as keys are always suffixes of the whole slices, the subslice length is enough information to uniquely determine the original subslice. Since the slices are short, we can then just jam both lengths into one u16. This speeds the program up dramatically, with the total runtime decreasing by ~80%.

Sadly, the cache is still sparse enough that using a dense array instead of a hashmap doesn't seem to be worth it. It might be possible to do some bit magic to intertwine the bits of the two lengths, so that the lower end of the array is more densely packed though.